Rename static-code-analysis example into decode opcodes
