+++ examples/decode-opcodes/README.md
@@ -1,0 +1,17 @@
+# Decode opcodes example
+
+This directory contains an example on how to decode opcodes
+
+The example does these things:
+
+1. Takes binary EVM code and decodes it into opcodes
+
+
+## Installation
+
+1. Run `npm install` in the root of this project
+
+## Running the example
+
+1. Run `npm run build:dist` in the root of this project
+1. Run `npm run example` in this directory
